requirements.txt
-i https://pypi.org/simple
aiofiles==22.1.0; python_version >= '3.7' and python_version < '4.0'
aiohttp==3.8.3
aiosignal==1.2.0; python_version >= '3.6'
async-timeout==4.0.2; python_version >= '3.6'
attrs==22.1.0; python_version >= '3.5'
charset-normalizer==2.1.1; python_version >= '3.6'
dnspython==2.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
frozenlist==1.3.1; python_version >= '3.7'
httptools==0.5.0; python_version >= '3.5'
idna==3.4; python_version >= '3.5'
jinja2==3.1.2
markupsafe==2.1.1; python_version >= '3.7'
motor==3.0.0
multidict==6.0.2; python_version >= '3.7'
natural==0.2.0
pymongo[srv]==4.2.0
python-dateutil==2.8.2
python-dotenv==0.18.0
sanic==22.6.2
sanic-routing==22.3.0
sanic-session==0.8.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ujson==5.5.0; sys_platform != 'win32' and implementation_name == 'cpython'
uvloop==0.17.0; sys_platform != 'win32' and implementation_name == 'cpython'
websockets==10.3; python_version >= '3.7'
yarl==1.8.1; python_version >= '3.7'
